code review
Your website is very minimalist style and clear. But It seems like I can't find a button to like/dislike each post, or that button is not obvious to me. In order to make your website more user-friendly, it would be great to have a button to clear all the filters and I can start all over again.